Add files via upload
Bug fixes. Added functionality to catch PRAW exceptions for Reddit "you are doing that too much" messages and turn them into a wait timer before retrying.

Changed how the city name is generated to deal with multiple cities with the same name in different states. Previously Burlington, VT and Burlington, NC would display as "BurlingtonVT" and "BurlingtonNC" in the post title and at the start of the post body.