Add files via upload
Fixed a bug that was causing the bot to populate the incorrect list of industries. It had been searching for a static bit of text to find the industries, but that text sometimes varies on the BLS website. I added functionality to identify the correct text and populate the industry list using the identified text.